+++ packages/leo-core/runtime-assets/load-leorc-from-develop.js
@@ -1,28 +1,39 @@
-var loadConfig = require('rc-loader');
+import cosmiconfig from 'cosmiconfig';
 var glob = require('glob');
 var path = require('path');
-var debug = require('debug')('leo:load-leorc-from-develop');
+var debug = require('debug')('leo:develop:load-leorc-from-develop');
 
 export default function loadConfig(callback) {
-  var conf = loadConfig('leo', {
-    rootPath: '/',
-    dataPath: './data',
-    extensions: ['.post', '.md']
-  });
+  cosmiconfig('leo', {})
+    .then(({ config, filepath }) => {
 
-  const exts = conf.extensions.join('|');
-  glob(`${conf.dataPath}/**/*`, {
-    nodir: true,
-    ignore: [
-      `${conf.dataPath}/**/*.jpeg`,
-      `${conf.dataPath}/**/*.jpg`,
-      `${conf.dataPath}/**/*.svg`,
-      `${conf.dataPath}/**/*.gif`,
-      `${conf.dataPath}/**/*.png`
-    ]
-  }, (err, files) => {
-    conf.files = files || [];
-    debug(`globbed ${files.length || 0} files`);
-    return callback(null, conf);
-  })
+      debug('using config found at', filepath);
+
+      // overwrite default values if set in userconfig
+      const conf = Object.assign({}, {
+        rootPath: '/',
+        dataPath: './data',
+        extensions: ['.post', '.md']
+      }, config);
+
+      const exts = conf.extensions.join('|');
+      glob(`${conf.dataPath}/**/*`, {
+        nodir: true,
+        ignore: [
+          `${conf.dataPath}/**/*.jpeg`,
+          `${conf.dataPath}/**/*.jpg`,
+          `${conf.dataPath}/**/*.svg`,
+          `${conf.dataPath}/**/*.gif`,
+          `${conf.dataPath}/**/*.png`
+        ]
+      }, (err, files) => {
+        conf.files = files || [];
+        debug(`globbed ${files.length || 0} files`);
+        return callback(null, conf);
+      })
+    })
+    .catch(parsingError => {
+      debug('parsing error', parsingError);
+      callback(parsingError);
+    });
 }
