+++ packages/leo-core/src/utils/enable-plugins.js
@@ -7,11 +7,14 @@
 import oDebug from 'debug';
 const debug = oDebug('leo:build:enable-plugins');
 
-export default function enablePlugins(plugins, config) {
-  plugins.forEach((plugin) => {
-    debug('enabling plugin', plugin);
-    // allow each plugin to add itself to the webpack config
-    require(plugin)(config);
+export default function enablePlugins(leoConf, config) {
+  leoConf.plugins.forEach((plugin) => {
+    debug('enabling plugin', plugin, 'with config', leoConf[plugin]);
+    /**
+     * allow each plugin to add itself to the webpack config
+     * with it's respective .leorc config, which may or may not exist
+     */
+    require(plugin)(config, leoConf[plugin]);
   })
     return config;
 }
